Anthony Brown | Wes Moore
The math was off before Anthony Brown and Wes Moore inherited office. I was affected in 2022, and millions of others years before that.

Now that we are all under sunlight these two can be the 
first ones to share the truth we deserve. And if they
share the truth and are retaliated against, the world will
know why. 

I won’t let them be scapegoated either. ( ⚠️ does not mean innocent)