Musical chairs in the Senate
Everyone is leaving. No mention of the people who are leaving. No mention of the hundreds that already left.

"If it was supposed to be musical chairs, it was a short song."

Rule # 1 of coverups, don't deny what no one accused.

"I did NOT... have pretextual relations... with those chairs."

*points finger*

message to the Maryland legislative branch - silence is a choice. You are being documented! 

"Everyone is doing it" is not cover. It just means everybody is complicit.

Oh and hire a new editor.